ai_agents_mcp/zerodha_mcp_agent/requirements.txt
kiteconnect==5.0.1
langchain_core==1.0.5
langchain_openai==1.0.3
langgraph==1.0.3
python-dotenv==1.2.1
streamlit==1.51.0
